chat-api/lambda-auth/auth-verify-clean/requirements.txt
google-auth==2.23.4
PyJWT==2.8.0
requests==2.32.5